pyproject.toml
[project]
name = "leverage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.55.0",
    "psycopg2-binary>=2.9.11",
    "scrapy>=2.13.3",
    "scrapy-playwright>=0.0.44",
]

[dependency-groups]
dev = ["pytest>=9.0.0", "ruff>=0.14.5"]
